Major changes for the daemon data transmission
This commit reengineer the data transmission from the session daemon to
the liblttngctl. Instead of sending a packet on a per data structure
basis, this scheme uses lttcomm_lttng_msg structure has a header control
structure adding the payload_size to it. The payload data is then
appended after and only one send() is used.

This will permit AF_INET socket scaling for future development.

List sessions and list apps command has been modified to fit this
change.

In the meantime, some unuse code is removed and minor cleanup.

+++ include/lttng/liblttngctl.h
@@ -46,7 +46,7 @@ extern int lttng_connect_sessiond(void);
 extern int lttng_set_tracing_group(const char *name);
 extern int lttng_check_session_daemon(void);
 extern const char *lttng_get_readable_code(int code);
-extern size_t lttng_ust_list_apps(pid_t **pids);
-extern size_t lttng_list_sessions(struct lttng_session **sessions);
+extern int lttng_ust_list_apps(pid_t **pids);
+extern int lttng_list_sessions(struct lttng_session **sessions);
 
 #endif /* _LIBLTTNGCTL_H */
